refactor: move posthog query and generic reponse to lib directory
