+++ packages/v1/quests/internal/queries/fetchdailyevents.go
@@ -2,16 +2,17 @@ package queries
 
 import (
 	"fmt"
+	"lib/types"
 	"lib/utils"
 	_types "quests/internal/types"
 )
 
 func FetchDailyEvents(account string, logger *utils.Logger) ([]_types.DailyEvent, error) {
-	var response _types.PostHogQueryResponse
+	var response types.PostHogQueryResponse
 	var result []_types.DailyEvent
 
 	query := fmt.Sprintf(`{"query":{"kind":"HogQLQuery","query":"select events.event as \"event-name\", count() > 0 as \"completed\" from events where person.pdi.distinct_id = '%s' and events.properties.genesisHash = 'IXnoWtviVVJW5LGivNFc0Dq14V3kqaXuK2u5OQrdVZo=' and events.timestamp >= today() group by \"event-name\" order by \"event-name\""}}`, account)
-	err := PostHogQuery(query, &response)
+	err := utils.SendPostHogQuery(query, &response)
 	if err != nil {
 		logger.Error(err)
